Module Vcs_param

Common command line parameters in use in vcs.

Initialization

module Config : sig ... end

val config : Config.t Command.Param.t

module Context : sig ... end

module Initialized : sig ... end

val initialize : env:< fs : _ Eio.Path.t ; process_mgr : _ Eio.Process.mgr.. > -> config:Config.t -> Initialized.t Base.Or_error.t

The initialization should be created very early in the commands body. The cwd shall not be changed subsequently.

Params

Some parameters may only be created under a certain context, in that case they're exposed wrapped under a resolvable type. Otherwise they can be exposed as command params directly.

If not otherwise specified, the parameters are resolved under the current context, and are required. Optional parameters ends with _optional.

type 'a t

A 'a Vcs_param.t is a command line parameter that produces a value of type 'a under the initialized context.

val resolve : 'a t -> context:Context.t -> 'a Base.Or_error.t

To be called in the body of the command, after initialization.

val anon_branch_name : Vcs.Branch_name.t Base.Or_error.t Command.Param.t

A required anon BRANCH.

val anon_branch_name_opt : Vcs.Branch_name.t Base.Or_error.t Base.option Command.Param.t

An optional anon BRANCH.

val anon_path : Fpath_base.Absolute_path.t t Command.Param.t

An anonymous parameter for a path. It can be given either as an absolute path or relative path in the command line, but will always be resolved to an absolute path.

val anon_path_in_repo : Vcs.Path_in_repo.t t Command.Param.t

An anonymous parameter for a path in repo.

val anon_rev : Vcs.Rev.t Base.Or_error.t Command.Param.t

A required anon REV.

val below_path_in_repo : Vcs.Path_in_repo.t Base.option t Command.Param.t

A flag to restrict the repo to a subdirectory below a certain path.

val commit_message : Vcs.Commit_message.t Base.Or_error.t Command.Param.t

A required -m _ nonempty commit message.

val quiet : Base.bool Command.Param.t

Perform the side effect if any, but suppress the output in case of success.

val rev : Vcs.Rev.t Base.Or_error.t Command.Param.t

A required --rev _ that produces a revision.

val user_email : Vcs.User_email.t Base.Or_error.t Command.Param.t

val user_name : Vcs.User_name.t Base.Or_error.t Command.Param.t